first-lcd.py
this is code for a 16x2 LCD screen used on a Raspberry Pi that cycles through the time, date, current temperature and humidity of the room, ip address of the device, and a Hello World. Now working on trying to include a buzzer if the temperature drops below a certain amount.